requirements.txt
dependency-injector>=4.39
boto3>=1.23
python-dotenv>=0.20